refactor: split
having a separate file which does not import anything from our fixtures
nicely works around this problem.
